fix(agent-gui): keep conversation projection DOM-free
