+++ packages/agent/gui/agent-gui/agentGuiNode/model/agentComposerDraft.ts
@@ -21,6 +21,13 @@ import {
   materializeAgentComposerFileMentions
 } from "./agentExternalPromptFiles";
 import { agentComposerFileMentionReferences } from "../agentRichText/agentMentionMarkdown";
+import { pastedTextDraftDisplayName } from "../../../shared/pastedTextReferenceProjection";
+
+export {
+  extractPastedTextArchivePaths,
+  linkifyPastedTextReferences,
+  pastedTextDraftDisplayName
+} from "../../../shared/pastedTextReferenceProjection";
 
 const PASTED_TEXT_MENTION_PREVIEW_MAX_CHARS = 10;
 
@@ -558,107 +565,6 @@ function escapeRegExp(value: string): string {
 export function textPromptContent(prompt: string): AgentPromptContentBlock[] {
   const text = prompt.trim();
   return text ? [{ type: "text", text }] : [];
-}
-
-/**
- * Display/label name for a pasted-text attachment, addressed purely by its
- * position in the draft (`pasted-text-1.txt`, `pasted-text-2.txt`, …). The
- * stored `item.name` is content-addressed and intentionally not used here, so
- * labels never collide and always renumber with the list.
- */
-export function pastedTextDraftDisplayName(index: number): string {
-  return `pasted-text-${index + 1}.txt`;
-}
-
-// Matches a landed pasted-text archive path (content-addressed .txt under the
-// host's agent-prompt-assets dir). The path may contain spaces (e.g. macOS
-// "Application Support"), so match from the leading "/" or drive letter up to
-// the first ".txt" after "agent-prompt-assets", staying on one line.
-const PASTED_TEXT_ARCHIVE_PATH_RE =
-  /(?:\/|[A-Za-z]:\\)[^\n]*?agent-prompt-assets[^\n]*?\.txt/;
-
-function firstPastedTextArchivePath(line: string): string | null {
-  return line.match(PASTED_TEXT_ARCHIVE_PATH_RE)?.[0].trim() ?? null;
-}
-
-/**
- * Extracts landed pasted-text archive paths from a persisted content text block
- * (the codex-style "Referenced pasted text files:" instruction the agent
- * receives). This is the reload-safe source of truth for the chip — see
- * {@link linkifyPastedTextReferences}.
- */
-export function extractPastedTextArchivePaths(text: string): string[] {
-  const paths: string[] = [];
-  for (const line of text.split("\n")) {
-    const path = firstPastedTextArchivePath(line);
-    if (path && !paths.includes(path)) {
-      paths.push(path);
-    }
-  }
-  return paths;
-}
-
-function pastedTextReferenceMentionMarkdown(
-  preview: string,
-  path: string,
-  index: number
-): string {
-  return createRichTextMentionMarkdown({
-    providerId: AGENT_PASTED_TEXT_MENTION_KIND,
-    entityId: `ref-${index}`,
-    label: preview.trim() || pastedTextDraftDisplayName(index),
-    scope: { path }
-  });
-}
-
-// The persisted instruction line embeds the preview quoted: `… "<preview>": …`.
-function firstQuotedPreview(line: string): string {
-  return line.match(/"([^"]*)"/)?.[1]?.trim() ?? "";
-}
-
-/**
- * Rewrites a persisted content text block that carries pasted-text references
- * into the same pasted-text mention chips the composer/display prompt produce,
- * so a reloaded message renders identical chips instead of the raw
- * "Referenced pasted text files: - pasted text file: <path>. Read this…" text.
- *
- * Mirrors the Codex approach (parse the agent-facing text back into attachment
- * chips): the pasted-text instruction is appended as its own content block, so a
- * block containing archive paths is entirely that section and is replaced by the
- * clean chip list (dropping the localized header/instruction wording). Blocks
- * without a pasted-text path are returned unchanged.
- */
-export function linkifyPastedTextReferences(text: string): string {
-  if (!PASTED_TEXT_ARCHIVE_PATH_RE.test(text)) {
-    return text;
-  }
-  const lines = text.split("\n");
-  const out: string[] = [];
-  let refIndex = 0;
-  for (let i = 0; i < lines.length; i++) {
-    const line = lines[i] ?? "";
-    const path = firstPastedTextArchivePath(line);
-    if (path) {
-      out.push(
-        pastedTextReferenceMentionMarkdown(
-          firstQuotedPreview(line),
-          path,
-          refIndex
-        )
-      );
-      refIndex += 1;
-      continue;
-    }
-    // Drop the localized header line that directly precedes a reference line;
-    // the pasted-text instruction always emits "<header>\n<refs>" as its own
-    // block, so this only removes the header, never user text.
-    const next = lines[i + 1];
-    if (next != null && firstPastedTextArchivePath(next)) {
-      continue;
-    }
-    out.push(line);
-  }
-  return out.join("\n").trim();
 }
 
 /**
